Intent: modularize metrics sections Operators involved: [Coherence, Observation, Self-organization] Affected invariants: [#7] Key changes: - extract glyph timing helpers to new module - move coherence/observer updates to dedicated module - relocate reporting functions and update import shims Expected risks/dissonances: minimal, relying on re-exports for backwards compatibility Metrics: N/A Equivalence map: tnfr.metrics.core API maintained via shims
